VideoMAEv2/requirements.txt
av
decord
deepspeed
einops
matplotlib
numpy
opencv-python
pandas
Pillow
scipy
tensorboard
tensorboardX
timm
torch
torchaudio
torchvision
triton
